Fix issue 183 - Prevent button delete from working after being cancel.
When user press cancel on confirm dialog after clicking a table delete
button, the ajaxec plugin is returning an initialized but empty data
object. Therefore, simply testing for an initialized object will
prevent calling ajaxec . We now also check that condition in order to
be able to call ajaxec plugin on an empty object as well.