Fix modal trigger close and few js fix
Now properly close modal. Problem was coming from jquery that was
include in final atk4js bundle. Need to change import $ from ‘jquery’
to const $ = require(‘jquery’) for it no to be include in bundle.
modal semantic is still include however. Even if set as external in
webpack config. need to find out why but does not seem to cause problem.