Fix plugin options error, style CI, add more modal example.
Plugin.js : needed to pass option to main function. Because calling the
plugin twice with different option on the same element was not working
since old option was still set in element.
Other plugin file: Need to adjust main function to accept options
spinner.js: add a function in order to remove spinner from element.
When calling the reloadView more than once on the same element the
element was growing because of spinner content not remove.
reloadView: adjust plugin to remove spinner on element being reload.