Added namespaces to quickstart and file structure example
I adapted and enhanced the things that have been missing for myself when I started using atk. So I first had to dig into namespaces and second I had to experiment with the file structure until I came up with one that is suited for small and big projects and offers a good overview and structure.