Add more examples to grid.php demo + issue changing id:1 record
I am adding further action buttons as a good example how to use a virtual page or addActionButton to attach customised forms to a grid.
There is a strange error that occurs: If you click the "envelope", the selected record's name will be lower cased - BUT the first record with id:1 will not. This seems to be a bug.